Refactor camp_application
