+++ config/locales/views.pbs.de.yml
@@ -50,30 +50,31 @@ de:
 
   export/pdf/camp_application:
     title: Lageranmeldung %{camp}
-    group_header: Abteilung und Einheit
-    leader_header: Lagerleiter*in
-    assistant_leaders_header: Lagerleitung
-    name: Name
-    address: Adresse
-    zip_town: PLZ / Ort
-    birthday: Geburtstag
-    qualifications: Ausbildung
+    by: durch
+  export/pdf/camp_application/group_section:
+    group_title: 1. Abteilung und Einheit
     abteilung: Abteilung
     einheit: Einheit
     expected_participants: Erwartete Anzahl Teilnehmende
-    camp: Lager
-    camp_dates: Lagerdaten
-    camp_location: Lagerort
-    j_s: J+S Rahmen
-    state: Lagerstatus
-    by: durch
-    abteilungsleitung: Abteilungsleitung
-    coach: Coach
     kantonalverband: Kantonalverband
+    leader_title: 2. Lagerleiter*in
+    qualifications: Ausbildung
+  export/pdf/camp_application/leaders_section:
+    title: 3. Lagerleitung
     assistant_leader_name: Vorname Name / Pfadiname
     assistant_leader_year: Jahrgang
     assistant_leader_qualifications: Gültige Ausbildungen
+  export/pdf/camp_application/camp_section:
+    title: 4. Lager
+    camp_dates: Lagerdaten
+    camp_location: Lagerort
+  export/pdf/camp_application/registration_section:
+    js_title: 5. J+S Rahmen
+    state_title: 6. Lagerstatus
     js_security: Sicherheitsbereich
+  export/pdf/camp_application/supervision_section:
+    abteilungsleitung_title: 7. Abteilungsleitung
+    coach_title: 8. Coach
 
   event:
     attendances:
